# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Volley Intelligence API",
    "description": "Org-scoped API for match management and pipeline processing.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/matches/{match_id}/sets": {
      "get": {
        "tags": [
          "ontology"
        ],
        "summary": "List Sets",
        "operationId": "list_sets_api_v1_matches__match_id__sets_get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchSetOut"
                  },
                  "title": "Response List Sets Api V1 Matches  Match Id  Sets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/{match_id}/rallies": {
      "get": {
        "tags": [
          "ontology"
        ],
        "summary": "List Rallies",
        "operationId": "list_rallies_api_v1_matches__match_id__rallies_get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RallyOut"
                  },
                  "title": "Response List Rallies Api V1 Matches  Match Id  Rallies Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rallies/{rally_id}/actions": {
      "get": {
        "tags": [
          "ontology"
        ],
        "summary": "List Rally Actions",
        "operationId": "list_rally_actions_api_v1_rallies__rally_id__actions_get",
        "parameters": [
          {
            "name": "rally_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rally Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActionOut"
                  },
                  "title": "Response List Rally Actions Api V1 Rallies  Rally Id  Actions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rallies/{rally_id}/analysis": {
      "get": {
        "tags": [
          "ontology"
        ],
        "summary": "Get Rally Analysis",
        "description": "Return the newest completed, internally consistent analysis bundle.\n\nThe result row carries ``organization_id`` specifically so the tenant\nboundary is enforced directly in the query. Joining the PipelineRun and\nrequiring COMPLETED prevents partially published worker output from\nbecoming visible in replay clients.",
        "operationId": "get_rally_analysis_api_v1_rallies__rally_id__analysis_get",
        "parameters": [
          {
            "name": "rally_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rally Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RallyAnalysisResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/{match_id}/statistics": {
      "get": {
        "tags": [
          "ontology"
        ],
        "summary": "Get Match Statistics",
        "description": "Computes statistics fresh from the Event Log every call -- per\nONTOLOGY.md, DerivedMetric is never persisted as a mutable number.",
        "operationId": "get_match_statistics_api_v1_matches__match_id__statistics_get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchStatisticsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/{match_id}/statistics/evidence": {
      "get": {
        "tags": [
          "ontology"
        ],
        "summary": "Get Stat Evidence",
        "description": "Return the exact Event Log rows behind one displayed statistic.\n\nThis replaces the browser's former one-request-per-rally fan-out with a\nsingle org-scoped query and a canonical server-side predicate shared by\nall consumers. The response is deliberately bounded and reports when it\nhas been truncated.",
        "operationId": "get_stat_evidence_api_v1_matches__match_id__statistics_evidence_get",
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/StatCategory"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Team Id"
            }
          },
          {
            "name": "zone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 6,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zone"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatEvidenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "phase_id": {
            "type": "string",
            "title": "Phase Id"
          },
          "rally_id": {
            "type": "string",
            "title": "Rally Id"
          },
          "action_type": {
            "type": "string",
            "enum": [
              "serve",
              "reception",
              "set",
              "attack",
              "tip",
              "block",
              "dig",
              "free_ball",
              "transition"
            ],
            "title": "Action Type"
          },
          "actor_team_id": {
            "type": "string",
            "title": "Actor Team Id"
          },
          "actor_roster_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Roster Id"
          },
          "video_t_start": {
            "type": "number",
            "title": "Video T Start"
          },
          "video_t_end": {
            "type": "number",
            "title": "Video T End"
          },
          "court_x": {
            "type": "number",
            "title": "Court X"
          },
          "court_y": {
            "type": "number",
            "title": "Court Y"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "reviewed_status": {
            "type": "string",
            "enum": [
              "unreviewed",
              "confirmed",
              "corrected"
            ],
            "title": "Reviewed Status"
          },
          "quality_rating": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Rating"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OutcomeOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "phase_id",
          "rally_id",
          "action_type",
          "actor_team_id",
          "actor_roster_id",
          "video_t_start",
          "video_t_end",
          "court_x",
          "court_y",
          "confidence",
          "reviewed_status",
          "quality_rating"
        ],
        "title": "ActionOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchSetOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "match_id": {
            "type": "string",
            "title": "Match Id"
          },
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "home_points": {
            "type": "integer",
            "title": "Home Points"
          },
          "away_points": {
            "type": "integer",
            "title": "Away Points"
          },
          "winner_team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Team Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "match_id",
          "index",
          "home_points",
          "away_points",
          "winner_team_id",
          "created_at"
        ],
        "title": "MatchSetOut"
      },
      "MatchStatisticsOut": {
        "properties": {
          "formula_version": {
            "type": "string",
            "title": "Formula Version"
          },
          "serve": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ServeStatsOut"
            },
            "type": "object",
            "title": "Serve"
          },
          "reception": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ReceptionStatsOut"
            },
            "type": "object",
            "title": "Reception"
          },
          "attack": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AttackStatsOut"
            },
            "type": "object",
            "title": "Attack"
          },
          "block": {
            "additionalProperties": {
              "$ref": "#/components/schemas/BlockStatsOut"
            },
            "type": "object",
            "title": "Block"
          },
          "dig": {
            "additionalProperties": {
              "$ref": "#/components/schemas/DigStatsOut"
            },
            "type": "object",
            "title": "Dig"
          },
          "sideout_breakpoint": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SideoutBreakpointStatsOut"
            },
            "type": "object",
            "title": "Sideout Breakpoint"
          },
          "setter_distribution": {
            "additionalProperties": {
              "$ref": "#/components/schemas/SetterDistributionEntryOut"
            },
            "type": "object",
            "title": "Setter Distribution"
          },
          "rally_duration": {
            "$ref": "#/components/schemas/RallyDurationStatsOut"
          }
        },
        "type": "object",
        "required": [
          "formula_version",
          "serve",
          "reception",
          "attack",
          "block",
          "dig",
          "sideout_breakpoint",
          "setter_distribution",
          "rally_duration"
        ],
        "title": "MatchStatisticsOut",
        "description": "Versioned, fully typed public contract for every computed metric.\n\nThe pure engine remains the formula source of truth; these models make\nshape changes visible in OpenAPI and therefore at TypeScript compile\ntime instead of failing inside a chart at runtime."
      },
...
      "OutcomeOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "result": {
            "type": "string",
            "enum": [
              "continue",
              "point",
              "error"
            ],
            "title": "Result"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "result",
          "detail"
        ],
        "title": "OutcomeOut"
      },
...
      "RallyAnalysisResultOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "match_id": {
            "type": "string",
            "title": "Match Id"
          },
          "rally_id": {
            "type": "string",
            "title": "Rally Id"
          },
          "video_id": {
            "type": "string",
            "title": "Video Id"
          },
          "pipeline_run_id": {
            "type": "string",
            "title": "Pipeline Run Id"
          },
          "schema_version": {
            "type": "string",
            "title": "Schema Version"
          },
          "content_sha256": {
            "type": "string",
            "pattern": "^[a-fA-F0-9]{64}$",
            "title": "Content Sha256"
          },
          "bundle": {
            "$ref": "#/components/schemas/RallyAnalysisBundle"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "match_id",
          "rally_id",
          "video_id",
          "pipeline_run_id",
          "schema_version",
          "content_sha256",
          "bundle",
          "created_at"
        ],
        "title": "RallyAnalysisResultOut",
        "description": "One immutable, completed professional analysis snapshot."
      },
...
      "RallyOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "set_id": {
            "type": "string",
            "title": "Set Id"
          },
          "index_in_set": {
            "type": "integer",
            "title": "Index In Set"
          },
          "serving_team_id": {
            "type": "string",
            "title": "Serving Team Id"
          },
          "point_winner_team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Point Winner Team Id"
          },
          "video_t_start": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video T Start"
          },
          "video_t_end": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video T End"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "set_id",
          "index_in_set",
          "serving_team_id",
          "point_winner_team_id",
          "video_t_start",
          "video_t_end",
          "duration_seconds"
        ],
        "title": "RallyOut"
      },
...
      "ServeStatsOut": {
        "properties": {
          "team_id": {
            "type": "string",
            "title": "Team Id"
          },
          "total_serves": {
            "type": "integer",
            "minimum": 0,
            "title": "Total Serves"
          },
          "aces": {
            "type": "integer",
            "minimum": 0,
            "title": "Aces"
          },
          "serve_errors": {
            "type": "integer",
            "minimum": 0,
            "title": "Serve Errors"
          },
          "zone_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Zone Counts"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "total_serves",
          "aces",
          "serve_errors",
          "zone_counts"
        ],
        "title": "ServeStatsOut"
      },
...
      "StatCategory": {
        "type": "string",
        "enum": [
          "serve_total",
          "serve_aces",
          "serve_errors",
          "reception_total",
          "attack_total",
          "attack_kills",
          "attack_errors",
          "block_total",
          "block_kills",
          "dig_total"
        ],
        "title": "StatCategory"
      },
...
      "StatEvidenceOut": {
        "properties": {
          "formula_version": {
            "type": "string",
            "title": "Formula Version"
          },
          "category": {
            "$ref": "#/components/schemas/StatCategory"
          },
          "team_id": {
            "type": "string",
            "title": "Team Id"
          },
          "zone": {
            "anyOf": [
              {
                "type": "integer",
                "enum": [
                  1,
                  2,
                  3,
                  4,
                  5,
                  6
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone"
          },
          "total_events": {
            "type": "integer",
            "minimum": 0,
            "title": "Total Events"
          },
          "returned_events": {
            "type": "integer",
            "minimum": 0,
            "title": "Returned Events"
          },
          "is_truncated": {
            "type": "boolean",
            "title": "Is Truncated"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/StatEvidenceEventOut"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "formula_version",
          "category",
          "team_id",
          "zone",
          "total_events",
          "returned_events",
          "is_truncated",
          "events"
        ],
        "title": "StatEvidenceOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}